refactor: Moved `DownloadPage` component code into Download Page
